TIP-842: Move attribute group repository to Akeneo\Pim\Structure (enrich bundle)
