+++ src/Akeneo/Pim/Structure/Bundle/Resources/config/repositories.yml
@@ -40,3 +40,26 @@ services:
         arguments: ['%pim_catalog.entity.attribute_option.class%']
         tags:
             - { name: 'pim_repository' }
+
+    ### External API
+    pim_api.repository.attribute:
+        class: 'Akeneo\Pim\Structure\Bundle\Doctrine\ORM\Repository\ExternalApi\AttributeRepository'
+        arguments:
+            - '@doctrine.orm.entity_manager'
+            - '%pim_catalog.entity.attribute.class%'
+            - '@pim_catalog.repository.attribute'
+
+    ### Intenal API
+      pim_enrich.repository.attribute_group:
+      class: '\Akeneo\Pim\Structure\Bundle\Doctrine\ORM\Repository\InternalApi\AttributeRepository'
+          arguments:
+              - '@pim_user.context.user'
+              - '@doctrine.orm.entity_manager'
+              - '%pim_catalog.entity.attribute_group.class%'
+
+    pim_enrich.repository.attribute:
+        class: 'Akeneo\Pim\Structure\Bundle\Doctrine\ORM\Repository\InternalApi\AttributeGroupRepository'
+        arguments:
+            - '@pim_user.context.user'
+            - '@doctrine.orm.entity_manager'
+            - '%pim_catalog.entity.attribute.class%'
