+++ src/Pim/Bundle/EnrichBundle/Resources/config/repositories.yml
@@ -10,13 +10,11 @@ parameters:
     pim_enrich.repository.group.class:                   Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\GroupRepository
     pim_enrich.repository.group_type.class:              Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\GroupTypeRepository
     pim_enrich.repository.association_type.class:        Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\AssociationTypeRepository
-    pim_enrich.repository.attribute.class:               Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\AttributeRepository
     pim_enrich.repository.family.class:                  Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\FamilyRepository
     pim_enrich.repository.family_variant.class:          Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\FamilyVariantRepository
     pim_enrich.repository.locale.class:                  Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\LocaleRepository
     pim_enrich.repository.channel.class:                 Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\ChannelRepository
     pim_enrich.repository.currency.class:                Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\CurrencyRepository
-    pim_enrich.repository.attribute_group.class:         Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\AttributeGroupRepository
     pim_enrich.repository.attribute_group.search.class:  Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\AttributeGroupSearchableRepository
     pim_enrich.repository.job_instance.class:            Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\JobInstanceRepository
     pim_enrich.repository.job_execution.class:           Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\JobExecutionRepository
@@ -71,13 +69,6 @@ services:
             - '@doctrine.orm.entity_manager'
             - '%pim_catalog.entity.association_type.class%'
 
-    pim_enrich.repository.attribute:
-        class: '%pim_enrich.repository.attribute.class%'
-        arguments:
-            - '@pim_user.context.user'
-            - '@doctrine.orm.entity_manager'
-            - '%pim_catalog.entity.attribute.class%'
-
     pim_enrich.repository.family:
         class: '%pim_enrich.repository.family.class%'
         arguments:
@@ -108,13 +99,6 @@ services:
         arguments:
             - '@doctrine.orm.entity_manager'
             - '%pim_catalog.entity.currency.class%'
-
-    pim_enrich.repository.attribute_group:
-        class: '%pim_enrich.repository.attribute_group.class%'
-        arguments:
-            - '@pim_user.context.user'
-            - '@doctrine.orm.entity_manager'
-            - '%pim_catalog.entity.attribute_group.class%'
 
     pim_enrich.repository.attribute_group.search:
         class: '%pim_enrich.repository.attribute_group.search.class%'
